+++ beatmap/v2/event.ts
@@ -128,7 +128,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if light  this is an off event.
      * ```ts
-     * if (isOff(event)) {}
+     * if (event.isOff()) {}
      * ```
      * ---
      * This may check non-light event too.
@@ -139,7 +139,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if light  this is an on event.
      * ```ts
-     * if (isOn(event)) {}
+     * if (event.isOn()) {}
      * ```
      * ---
      * This may check non-light event too.
@@ -150,7 +150,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if light  this is a flash event.
      * ```ts
-     * if (isFlash(event)) {}
+     * if (event.isFlash()) {}
      * ```
      * ---
      * This may check non-light event too.
@@ -161,7 +161,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if light  this is a fade event.
      * ```ts
-     * if (isFade(event)) {}
+     * if (event.isFade()) {}
      * ```
      * ---
      * This may check non-light event too.
@@ -172,7 +172,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if light  this is a transition event.
      * ```ts
-     * if (isTransition(event)) {}
+     * if (event.isTransition()) {}
      * ```
      * This may check non-light event too.
      */
@@ -180,9 +180,42 @@ export class Event extends BeatmapObject<IEvent> {
         return this.value === 4 || this.value === 8 || this.value === 12;
     };
 
+    /** Check if light event is a red light.
+     * ```ts
+     * if (event.isRed()) {}
+     * ```
+     * ---
+     * This may check non-light event too.
+     */
+    isRed(): boolean {
+        return this.value === 5 || this.value === 6 || this.value === 7 || this.value === 8;
+    }
+
+    /** Check if light event is a blue light.
+     * ```ts
+     * if (event.isBlue()) {}
+     * ```
+     * ---
+     * This may check non-light event too.
+     */
+    isBlue(): boolean {
+        return this.value === 1 || this.value === 2 || this.value === 3 || this.value === 4;
+    }
+
+    /** Check if light event is a white light.
+     * ```ts
+     * if (event.isWhite()) {}
+     * ```
+     * ---
+     * This may check non-light event too.
+     */
+    isWhite(): boolean {
+        return this.value === 9 || this.value === 10 || this.value === 11 || this.value === 12;
+    }
+
     /** Check if  this is a valid type.
      * ```ts
-     * if (isValidType(event)) {}
+     * if (event.isValidType()) {}
      * ```
      */
     isValidType = (): boolean => {
@@ -191,7 +224,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a light event.
      * ```ts
-     * if (isLightEvent(event)) {}
+     * if (event.isLightEvent()) {}
      * ```
      */
     isLightEvent = (): this is EventLight => {
@@ -210,7 +243,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a boost event.
      * ```ts
-     * if (isColorBoost(event)) {}
+     * if (event.isColorBoost()) {}
      * ```
      */
     isColorBoost = (): boolean => {
@@ -219,7 +252,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a ring event.
      * ```ts
-     * if (isRingEvent(event)) {}
+     * if (event.isRingEvent()) {}
      * ```
      * ---
      * This does not check for ring zoom.
@@ -230,7 +263,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a laser rotation event.
      * ```ts
-     * if (isLaserRotationEvent(event)) {}
+     * if (event.isLaserRotationEvent()) {}
      * ```
      */
     isLaserRotationEvent = (): this is EventLaser => {
@@ -239,7 +272,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a lane rotation event.
      * ```ts
-     * if (isLaneRotationEvent(event)) {}
+     * if (event.isLaneRotationEvent()) {}
      * ```
      */
     isLaneRotationEvent = (): this is EventLaneRotation => {
@@ -248,7 +281,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a extra event.
      * ```ts
-     * if (isExtraEvent(event)) {}
+     * if (event.isExtraEvent()) {}
      * ```
      */
     isExtraEvent = (): boolean => {
@@ -257,7 +290,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a special event.
      * ```ts
-     * if (isSpecialEvent(event)) {}
+     * if (event.isSpecialEvent()) {}
      * ```
      */
     isSpecialEvent = (): boolean => {
@@ -266,7 +299,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a BPM change event.
      * ```ts
-     * if (isBPMChangeEvent(event)) {}
+     * if (event.isBPMChangeEvent()) {}
      * ```
      */
     isBPMChangeEvent = (): boolean => {
@@ -275,7 +308,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Not to be confused with isLightEvent, this checks for event that affects the environment/lighting.
      * ```ts
-     * if (isLightingEvent(event)) {}
+     * if (event.isLightingEvent()) {}
      * ```
      */
     isLightingEvent = (): boolean => {
@@ -284,7 +317,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if event has Chroma properties.
      * ```ts
-     * if (hasChroma(event)) {}
+     * if (event.hasChroma()) {}
      * ```
      */
     // holy shit i hate type guard
@@ -328,7 +361,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if event has old Chroma properties.
      * ```ts
-     * if (hasOldChroma(event)) {}
+     * if (event.hasOldChroma()) {}
      * ```
      */
     hasOldChroma = (): boolean => {
@@ -337,7 +370,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if event has Noodle Extensions properties.
      * ```ts
-     * if (hasNoodleExtensions(event)) {}
+     * if (event.hasNoodleExtensions()) {}
      * ```
      */
     hasNoodleExtensions = (): boolean => {
@@ -346,7 +379,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if event has Mapping Extensions properties.
      * ```ts
-     * if (hasMappingExtensions(event)) {}
+     * if (event.hasMappingExtensions()) {}
      * ```
      */
     hasMappingExtensions = (): boolean => {
@@ -355,7 +388,7 @@ export class Event extends BeatmapObject<IEvent> {
 
     /** Check if  this is a valid, vanilla event.
      * ```ts
-     * if (isValid(event)) {}
+     * if (event.isValid()) {}
      * ```
      */
     isValid = (): boolean => {