+++ extensions/chroma/environment/mod.ts
@@ -1,3 +1,4 @@
 export * from './block.ts';
-export * from './group.ts';
 export * from './converter.ts';
+export * from './grab.ts';
+export * from './group.ts';